conformance with 64 bit mode compilation
Less strict acceleration times check: 4 instead of 5 times

+++ src/cspec/cspec_output_junit_xml.c
@@ -98,7 +98,7 @@ int sumup_failure(const descrOutputs_t* const descr)
     int sum = 0;
 
     for (j = 0; j < descr->n_itOutputs; ++j) {
-        sum += descr->itOutputs[j].failures->size;
+        sum += (int)descr->itOutputs[j].failures->size;
     }
     return sum;
 }
